fix(indexing): a step longer than five minutes was killed for being slow (N1)
`RunCoordinator.step` wrote `heartbeat_at` on entry and again on success, and nothing
between. `IndexingRun` is the row `StaleRunReaper` reaps, so any single step outliving
`stale_running_heartbeat_timeout_seconds` (300) was declared dead *while it was still
working* and flipped to failed with `error='stale run reaped'`.

Production had failed this way daily since 2026-08-07: 64 of 70 repo-index failures,
all `current_step='graph_build'`. Runs made of many short steps completed at 417 s and
2 568 s — the signature of a per-step limit, not a per-run one.

It hid for thirteen days because a heartbeat already existed on the wrong row.
`_run_index_background` (repos.py:556-563) ticks `IndexingCheckpoint`, so the checkpoint
stayed alive beside a run being reaped, and nothing looked like a missing heartbeat.

- the beat wraps the step body, in `step` rather than in a caller: all four repo-index
  entry points (ARQ task, manual route, retry route, daily sync) reach the pipeline
  through this contextmanager, and `IndexingRun` is created and finished here. Wrapping
  `worker.py`'s `run_repo_index` — the first fix considered — covers one of the four and
  puts a second heartbeat beside the existing one
- the writer opens its own session; the step's `db` driven from two tasks is a race,
  not a heartbeat
- targeted UPDATE, not an ORM load: it must not carry a stale `version` into a column
  the step boundaries bump deliberately, nor win on columns it does not own
- `heartbeat()`'s interval is now `float`. The body always did float math; the `int`
  annotation rounded a sub-second interval to zero, so no test could stand in for a
  300 s timeout without waiting 300 s

Three tests. The third asserts the beat STOPS when the step does — a heartbeat that
outlives its step keeps a crashed run immortal, which is the failure the reaper exists
to catch, reintroduced by its own fix. All three verified red before the change.

Suite: 6747 passed, 4 skipped, 1 xfailed (was 6744 — the three new ones).

+++ backend/app/core/heartbeat.py
@@ -32,8 +32,11 @@ async def _beat(writer: HeartbeatWriter, interval_seconds: float) -> None:
 async def heartbeat(
     writer: HeartbeatWriter,
     *,
-    interval_seconds: int,
+    interval_seconds: float,
 ) -> AsyncIterator[None]:
+    # float, not int: the body already does float math (`max(0.001, …)`), and a
+    # caller that wants a sub-second beat — a test standing in for a 300 s timeout —
+    # had its interval rounded to zero by the annotation's implied contract.
     # Immediate first beat so a just-started row gets heartbeat_at before the
     # first interval elapses.
     try:
